move LayerFreezingHparams
